done subject, user, teaching

+++ Backend/src/main/java/com/example/net/guides/springboot/repository/UserRepository.java
@@ -1,12 +1,12 @@
 package com.example.net.guides.springboot.repository;
 
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.net.guides.springboot.model.User;
-@Repository 
-public interface UserRepository extends JpaRepository<User, Long> {
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findByUsername(String username);
 }
